Exploratory Data Analysis - Terrorism Project
This Python project focuses on Exploratory Data Analysis of global terrorism data. It involves cleaning, transforming, and visualizing the dataset to gain insights into patterns and trends related to terrorist attacks from 1970 to 2017. The analysis includes visualizations of attack frequency over time, casualties, and other statistics to provide an understanding of the impact and nature of these attacks worldwide